WL#15116 Remove the memcached plugin
The memcached library bundled with MySQL has been deprecated in 8.0 .
The library is removed in this patch.

This patch removes:
   the plugin/innodb_memcached source directory,
   any CMake reference to it (hence WITH_INNODB_MEMCACHED is not
      a known option anymore),
   all special code paths, workarounds under storage/innodb, sql/
      that were marked to exist specifically for purpose of
      supporting the memcached InnoDB plugin, including parts
      of the InnoDB API
   all tests that used memcached

Some parts of the InnoDB API that are not used anymore are not
removed, as determining which parts were specifically only used
by Memcached is made difficult by InnoDB itself using calls to
the API internally. This task is left up to a more general
refactoring -- since there is no other external (external to InnoDB)
user of the API, it should probably not be called an API.

Change-Id: Ia60d77d2e898e63d79afae89ddc2c7a09bdb18b8