Bug#35882307 use-after-free at disconnect request
Running the router's tests with high concurrency and low CPU count

  taskset -c 0-1 ctest --output-on-failure -R routertest_ -j 8

reports random test-failures with stacktraces.

The stacktraces hint at a use-after-free of the connection object
after a "async_send() -> async_recv()" transition.

Background
==========

The use-after-free is triggered by:

1. a disconnect-request (via a shutdown signal like SIGTERM) right
   BEFORE ...
2. router starts to wait for both client and server to send something.

With a normal test, the disconnect-request (at test-end) happens AFTER
the router started to wait for both client and server to send the next
command.

The underlying issue is that in case of "RecvFromBoth" the
"async_recv_server()" is called even when the "async_recv_client()"
closed the connection on a disconnect-request:

- disconnect-request
- async_recv_client()
  -> check if disconnect was requested?
  -> yes, finish() -> close socket -> remove connection()
- async_recv_server() -- use-after-free

Change
======

- add a async_recv_both() function which handles the disconnect-request
  properly.
- move the "recv-from-both" handling from async_recv_server() and
  async_recv_client() into async_recv_both()

Change-Id: I241d48e5cb291ac1e21fd8a48fb5b6609a962af7