Bug#35861516 Ndb : DbUtil test class only supports prepared statements
By default all SqlClient(formerly known as DbUtil) queries are ran using
the prepared statement interface. This is done in order to support
placeholders (i.e '?') in the query to be replaced with provided
arguments. Since there are very few tests which takes advantage of this
functionality a new function using the basic interface is added and
invoked for all queries where no arguments are provided.

This makes the SqlClient support running all possible SQL queries.

Change-Id: Idd77098221cec75c90f39962d8031b7fa17a0f72