Bug#35889188 routertest_component_gr_state ClusterSetAccessToPartitionWithNoQuorum fails
Some of the tests in routertest_component_gr_state do not wait for Router READY
notification (as it is not expected in those). Instead they assume that waiting
for the metadata-cache plugin to begin queries is enough for the Router to start
the signal handler. This is not always the case and sometimes the Router is
missing the shutdown singal causing the test to fail.

This patch uses RUNNING notification for those tests making sure the singal
handler is always up and running before the test tries to shut down the Router.

Change-Id: I23212daf4f412cad42c4fe28e4b555a3b555f34b