Bug#35853405 openssl-based remote key signing fails
When the --remote-openssl option is used in ndb_sign_keys,
--CA-cert should be set to the full pathname of the
certificate on the remote server, and --CA-key should be
set to the full pathname to the key file. In every
other case besides --remote-openssl, --CA-cert and --CA-key
simply hold filenames, and the TLS search path is used.
The exception is a bit unexpected and should be documented.

In ndb_sign_keys:
  When using --remote-openssl, the CA cert must be fetched
    from the remote server.
  Fix a logic error checking opt_remote_path.
  Fix an error exit that required local CA credentials.
  Fix several memory leaks.
  Fix the set of options supplied to the remote openssl command.

In testMgmd, extend the SshKeySigning test to also test signing
with --remote-openssl, and add a few tests to check the resulting
certificates.

Change-Id: I0b61b102cd97a0c9f3c0b24d12a4955c59d44e18