BUG#35893697 - Unable to access REST database object with a Web Browser
Description
===========
The format of the "Accept" header allows additional parameter appended
to the MIME type. The full format which should be supported is following:

Accept: mime-class/mime-subclass[;q=NUMBER][,mime-class/mime-subclass[;q=NUMBER]]+

The weighting parameter ("q") is treated as part of "mime-subclass".

Fix
===
Extract and ignore the weighting parameter.