Bug#35710373: mysql server 8.1.0 failed at Item::has_aggregation
The problem is that Item_singlerow_subselect::collect_scalar_subqueries
is called with a row subquery when m_collect_unconditionally is true.
The function is not prepared for this and thus there is a failure.

The fix is to give a return with a false value when called with
a row subquery.

Change-Id: I7eeb6580ceb6f27dba0b2cb18998e76e63ff416b