Bug#35082638: Report the cause of failure of monitor thread in managed replication channel
Issue:
======
The Monitor IO thread for a managed replication channel (channel with
asynchronous connection failover mechanism enabled) always logs a
generic error message in the error log for failure in executing query
without specifying exact reason of failure. When performance_schema
table access is denied to Monitor IO thread it only logs 'The Monitor
IO thread failed to detect if the source belongs to the group majority
on the source', which is incomplete error message. Although, the query
was executed to find if the source belongs to the group majority, but
query failed because of missing privilges which was not added in the
error log.

Solution:
=========
Added more details to the error message, for the monitor and replica IO
thread while executing query on the source server.

Change-Id: Ib4bf3d399bb4a7e80d1b09fc433c2cc64a2a2544