FEATURE: File sharing using service/db_schema/content_set json-options
User may define entries in the json-options to define files, directory-index,
redirections on service, schema, content set.

Here is the supported format:
{
    "defaultStaticContent": {
        "index.html": "PCFET0NUW...",
        "favicon.ico": "AAABAAMAM...",
        "favicon.svg": "PD94bWwmV..."
    },
    "defaultRedirects": {
        "redirect.html": "/myService/myContentSet/index.html"
    },
    "directoryIndexDirective": [
        "index.html"
    ]
}

When set on service /srv10, then is going to generate following files:

* /srv10/index.html
* /srv10/favicon.ico
* /srv10/favicon.svg
* /srv10/redirect.html
* /srv10/  -> maps to /srv10/index.html
* /srv10   -> maps to /srv10/index.html

The behaviour is corresponding for schema and content set.

Authentication settings are derived from the object on which it was set (schema, content_set).
The file content may be set as raw text value of base64 encoded text.