Bug#35861516 Ndb : DbUtil test class only supports prepared statements
Follow up fix for problem with AtrtClient which attempted to read a
number which had been saved only as a string.

Improve test case to check that both prepared and basic protcol returns
same values and that they are save in same format where expected.

In general the basic protocol returns all values as strings which
differs against the prepared statement protocol, thus the mappings are
not always identical  but it seems like the basic protocol currently
have a more resilient implementation.

Change-Id: Idf36139e4cf82d93cc9215a63ad516b8ffd76887