Bug#35878975: mysqld crash - Assertion failure in OverflowBitset::Or at overflow_bitset.h
An assertion failure in seen when executing certain queries with a
degenerate outer join condition.

There is some code in the hypergraph optimizer that massages the
access path tree if it contains certain degenerate join conditions, so
that HeatWave is able to cope with them. In particular, HeatWave does
not accept left outer joins if the join condition references tables on
the right side only. The hypergraph optimizer usually tries to push
such conditions further down, but sometimes it doesn't have anywhere
to push them, and leaves them as the join conditions they were
originally specified as. A workaround for HeatWave detects these
conditions during planning, and manually moves them from the join
condition to a filter just below the join.

The workaround failed to copy some bit sets from the original access
path to the injected filter access paths. Those bit sets are usually
empty when the workaround is applied, but when they are not copied,
the filter path ends up having empty inlined OverflowBitSets even if
the number of predicates in the query exceeds the capacity of inlined
bit sets (63). This leads to assertion failures later in the planning
when the optimizer expects the bit sets to have big enough capacity to
hold one bit per predicate in the query.

Fixed by copying the bit sets from the original path to the filter
path.

Change-Id: I064df2c5ba6d40664877043d720bef41a4f1ccde