Bug#35750394 Remove theNodeIdTransporter[] state invariant
Refactoring patch removing invariant where there
could be a MultiTransporter created, but not set up
as theNodeIdTransporter[] yet.

1) When a data node is restarted, the MultiTransporter is
created already when reading the config. However it is not
inserted into the theNodeIdTransporter[] until after
QMGR has 'switched' to use the the MultiTransporter.

2) When we are reconnecting to a data node where a MultiTransporter
were previously used, it already existed in theNodeIdTransporter[],
even if the reconnect process had not yet reached the 'switch'
point.

At the same code in QMGR, TRPMAN and TransporterRegistry is handling
both the intial connect and reconnect, it is complicated by
that a connecting MultiTransporter may or may not exist in
theNodeIdTransporter[] when setting up such a connection.

This patch insert the MultiTransporter into theNodeIdTransporter[]
as soon as it is created. Note that if the QMGR handshaking with
neighbour node will then later conclude that multiple transporters can
not be used, we will simply continue to communicate over the
single base transporter which is always connected first.

This allows us to simplify get_node_multi_transporter()
and remove a couple of other get'ters for multi_transporter.

Change-Id: I152b3a2821e0e1eac087f6b864fa97962f98dacf