Bug#35750394 Refactor Transporter interface to use TrpId over NodeId
Refactoring patch required to allow the transporter
CONNECTING/DISCONNECTING protocol to be used for the
individual PartsOfMultiTransporter's.

The usage of a 'NodeId' to identify a Transporter in the interface
is replaced by using its 'TrpId' instead. NodeBitmask is also
replaced by TrpBitmask.

Naming of methods now taking modified arguments are mostly
kept unchanged. The execpetion is where variants taking
both a NodeId and TrpId are needed, where the later then is
named with a *_trp suffix.

Format of EventLogger messages as well as different
DEBUG output has been addapted such that ut should be
clear whether a NodeId or a TrpId is reported in the logs.
(Somethimes both as well)

Some comments starting with 'OJA' are added by the patch,
pointing out where MultiTransporters does not follow the
correct protocol to disconnect or connect. These comments
will be removed again by the following patches which are
the real fixes.

Change-Id: I02ab4daa46e4b19d162c234bf4377cf3e918372a