Bug#35750394 Rename connect/disconnect methods:
'reconnect' and 'disconnect' seems overused in naming
methods and the nameing might not always help to clearify
the intended usage of each method. In particular some
of these methods do an immediate connect and waits for
the connection with the other peer to be
established/teared down, while others starts the asynch
CONNECTING/DISCONNECING protocol.

Some of the root cause for this bug might be that the
naming of the methods doing the different steps in connect/
disconnect of transporters were confusing. Many of them were
named 'do connect', with different variants of upper case and '_'
in their naming, and some simply 'connect'. In these cases
'do' might be misleading and indicate an immediate
connect/disconnect taking place.

That made it unclear wheteher the particular method did an
immediate connect or just initated the CONNECTING/DISCONNECTING
transporter protocol which would eventually result in the transporter
being reported as CONNECTED/DISCONNECTED, but relied on the
start_clients_threads to do the real connect/disconnect, and the
final CONNECTED/DISCONNECTED state to be set by report_connect/
report_disconnect.

This patch rename such 'do connect/disconnect' methods to
'start connecting/disconnecting', in particular:

  TransporterRegistry::do_connect        -> start_connecting
  TransporterRegistry::do_connect_trp    -> start_connecting_trp

  TransporterRegistry::do_disconnect     -> start_disconnecting
  TransporterRegistry::do_disconnect_trp -> start_disconnecting_trp

  Transporter::do_disconnect             -> start_disconnecting

  TransporterFacade::doConnect           -> startConnecting
  TransporterFacade::doDisconnect        -> startDisonnecting

Change-Id: I02577788ff66cfaee50f30019ad18329ec1f246e