Bug#35750394 Use transporter protocol to connect 'PartsOfMultiTransporter'
We used to connect the PartsOfMultiTransporter when
start_clients_thread discovered that the the base
transporter had CONNECTED by calling
Transporter::connect_client() directly.

It basically did some of the same steps as if the
transporters were in a CONNECTING state. However,
as the CONNECTING-protocol was not used, all the
different state transitions, enable_send_buffers,
epoll_add and such had to be patched up as an
afterthought by QMGR.

Generally the protocol to set up connections were not
followed either, were there are strict rules of where the
performState[] transitions can be done, E.g. the CONNECTED
state should only be set by the receive thread responsible
for the specific TrpId.

Patch is to let Qmgr::connect_multi_transporter() use
start_connecting_trp(TrpId) to initate the CONNECTING
protocol. It will the connect as any other
transporters.

Also note that the 'hack' to set_s_port() on the
transporter now is obsolete - will be done by
start_clients_thread when seeing the CONNECTING
transporter.

Removed epoll_add functions and other code now obsolete.

Change-Id: I8937d96a2a1b7440a84b0dc3a0c8d2b857b7a2b7