Bug#35750394 Use transporter protocol to disconnect 'PartsOfMultiTransporter'
We used to disconnect the MultiTransporters when report_disconnect
of *one* of the MultiTransporters was observed by calling
Transporter::forceUnsafeDisconnect(). Even if the disconnect
being 'reported' followed the disconnect protocoll, the other
disconnects of the remaing transporters *did not*. Besides
the protocol breakage, this could also cause  real time execution
problems as closing of sockets could stall, waiting for the
other peer or timeouts.

Patch fixes this by starting an asynchronous disconnect
instead with start_disconnecting_trp().

Transporter::forceUnsafeDisconnect is now obsolete
and is removed.

Transporter::start_disconnecting can be simplified
as there should be no difference between a disconnecting
multiTransporter and a plain transporter any more.

Change-Id: I3101603a66b7f712d2e51f2c9c0ae963d67edddb