Bug#35750394 Replace shutdown() with start_disconnecting_trp()
When QMGR has concluded that all the PartsOfMultiTransporter has
CONNECTED, it will disconnect the base transporter which is
not needed anymore.

This used the Transporter::shutdown() method which forcefully just
closed theSocket, not following the transporter protocol of first
setting the performState[] to DISCONNECTING, and letting the asynch
disconnect happening out of the way of the executing block / QMGR.

Patch replace usage of ::shutdown() with start_disconnecting_trp().
::shutdown becomes obsolete and is removed.

Change-Id: Iee22f2c48df0e650fd29e5e3d12d6f0d06b5059b