BUG#35918849 Automatic thread_end in SqlClient
When the SqlClient uses the MySQL library it allocate resources which
need to be released using mysql_thread_end(). Such resource leakage is
visible with asan or valgrind.

This is fixed by changing SqlClient to activate a thread_local guard to
release the MySQL thread resources when thread ends.

Also make sure to only initialize the MySQL C API library once, this is
important when running with multiple threads.

Change-Id: I6f49b2725c725c809f3cacb1300619e078e449c7