Bug#34943382 	mysqltest: At line 3: Unexpected result 4294967295 from WaitForSingleObject
mysqltest can execute `Q_SKIP` command and call `cleanup_and_exit()` before the `wait_stacktrace_request_event()` threads start listening for the event the `cleanup_and_exit()` cleans up.
The fix is to not close this event from the main thread, but only from the `wait_stacktrace_request_event()` thread. For this we need to just signal it, but we don't want it to call `handle_wait_stacktrace_request_event` so we wrap it with check for a newly added flag to distinguish event signal from `cleanup_and_exit()` from others.
Additionally, the collector thread can itself call the `die()` method, in which case the use-after-free, buffer underrun and other problems happen because of wrong memory management, especially in the `free_used_memory()` method or in `close_files()` where the `cur_file` is set to element before the `file_stack` array starts - and it is being used later for dereferencing.

Change-Id: I6dcaa8cce7b913906cff1b9e2c48b9eb679c954b