Bug#35923771	binlog_transaction_compression may silence ER_RPL_MTA_STATISTICS messages
Issue description
-----------------

When a source MySQL server has binlog_transaction_compression = ON, DML workloads
then to be written into the binary log as a sequence of two transactions:
the GTID and the PAYLOAD. A replica applying such workload may not report
ER_RPL_MTA_STATISTICS for a long time because of the logic used to decide
if the statistics message shall be generated or not.

Analysis
--------
Message ER_RPL_MTA_STATISTICS is written every time number of added events
executed by coordinator modulo 1024 equals 1. When even number of events
is added then the condition never holds.

Proposed solution
-----------------
Code which is executed every 1024 events needs to be changed so that
it is execute every over 1024 events.

Change-Id: I317b55b47d9466dc48585696c9169205b37246c3