Fix startup hang when memcached enabled and DB has many tables
As reported in https://bugs.mysql.com/bug.php?id=103275 ,
when memcached is enabled and a MySQL DB has many tables,
the database hangs on startup due to a deadlock
between the memcached plugin and the mysqld main thread.
This deadlock is because memcached is waiting for
a lock that is held by the main thread whereas the main
thread is waiting on a condition variable whose notifier
is unknown.

The cause of this deadlock is a race condition
created at startup when the mysqld main thread is resetting
the shared dictionary cache while at the same time memcached
is populating this same cache as it accesses underlying
objects during its own initialization.

To solve the problem, we ensure ensure strict ordering
between shared dictionary cache resetting and memcached
initialization. We use a single flag protected in a
critical region to communicate completion of shared
dictionary cache resetting between the mysqld main thread
and the memcached thread. The memcached polls this flag
with a back-off period of 1 second in case it is not set. It
is expected that the back-off period never to be hit more than
once as the resetting of the shared dictionary cache is a
quick process, and so if the memcached thread polls the flag
once and finds it unset, it is likely to find it set on
the next try.

Rationale for implementation
============================
The mutex lock has got 2 reasons for its usage in
the implementation:

- To ensure that access to flag are atomic: We do not want
a scenario in which the writer thread partially
updates the flag and then gets evicted from execution by the
OS, and then the reader thread reads the flag while it is in
an inconsitent state. This scenario in modern C++ is normally
handled by std::atomic_bool (std::atomic<bool>), but is not
available in C for use in memcached.
- We want to ensure that there is no compiler or CPU reordering
of instructions on either side of the pthread_mutex. Normally
with C++ one would use std::atomic_thread_fence, but it is not
usable for this case because memcached is a C library rather
than a C++ one. When we attempted to use it, we got linker
errors, so we decided to create a "poor man's" fence.
pthread_mutex has fencing capabilities, see
https://stackoverflow.com/a/24138417 , and so using it allows
us to make sure that the dd cache clearing instructions
fully run and set the flag, before the memcached
initialization happens hence ensuring strict ordering.

Reason for implementing a polling with back-off period
in the memcached plugin thread:

We want to avoid any additional synchronization mechanisms
as our main goal is ordering and not multithreaded
synchronization, so we opt not to use pthread_cond_timedwait
but instead do the following:

- If we successfully lock the mutex, we check the flag, and
if it is set, we just move on to initialization of the
memcached plugin.
- If we successfully lock the mutex, but find that the flag
is not set, then we want to signal to the OS that we want the
memcached thread to be suspended so that other threads can
take its execution slot. Even though whichever thread that is
chosen to take the memcached thread's slot is random, we are
at least giving a chance for the main thread to to proceed
and hopefully complete dd shared dictionary cache resetting.
- If we do not get the lock, we want to signal to the OS
that the memached thread's execution slot can be ceded to
another thread while we wait for 1 second. During this
time we are expecting the resetting of the dd shared
dictionary cache will have completed. I.e. we expect this
back-off to only happen once.

This contribution is under the OCA signed by Amazon and covering
submissions to the MySQL project.