Bug#35907828 testNodeRestart -n LCP_with_many_parts_drop_table  fails when Partial LCP is OFF
Problem:
LCP_with_many_parts_drop_table is intended to cover a scenario where
an LCP is restored from ~2048 partial LCPs with 1 part each one.
To increase the number of partial-LCPs needed to restore an LCP the
ERROR insert 10048 is used to force each one to use only 1 part.
This works when partial LCP is enabled but for full LCP the error
insert forces the number of parts calculated to be 1 instead of
2048 (default).
This causes the assertion:
ndbassert(is_partial_lcp_enabled() || num_change_parts == 0)
to fail due to the wrong number of parts.

Solution:
Disable LCP_with_many_parts_drop_tables when EnablePartialLcp = 0
since test is useless in these case.
Also, to prevent future issues caused by EI 10048 when partial LCP
is disabled, error insert 10048 handler changed to set the number
of parts to 1 only when partial lcp is enabled. For full lcp EI 10048
does nothing.

Change-Id: I7cfa9628d371602761f600fa7a3b2d01cd61cf18