Bug#35790567 redo_log_recovery_state test is failing for page-size=4K
Description
===========
This test was added recently through Bug#34687854. Aim of the test was
to validate the FIL_PAGE_LSN is updated correctly in the pages after
recovery. Test compared the two states of the tablespace file. The
comparison was failing for the page-size=4K.

As reported in the Bug#35912429, it may happen that innodb updates the
page LSN because it adds the non-dirtied but S[X] latched pages to the
flush list as well. Therefore, the page LSN of a page may be updated
even though page data was not changed.

In this case the test failed because the page lsn on the first lob page
(that is page#5) differed between two states of the tablespace.

If you are wondering "Why page#5?" then you may like to read excellent
blog titled :
"MySQL 8.0: InnoDB Introduces LOB Index For Faster Updates" from Anna.

If we inspect the lob::insert() method, we notice that method restarts
the btree mtr on the every fourth page written because commit_freq
variable is set to 4, and it re-acquires the X latch on the first lob
page. If there is no data remaining(=0) to be written then mtr is
committed by the caller. The log buffer of the mtr contains 32 byte
long redo log record describing change to page#4, but no changes
to the page#5. Still, both page#4 and page#5 are referenced from
m_impl.m_memo as being X-latched by this mtr, so both are added to the
flush lists.

Thus when flushed, page#5 of the tablespace has page lsn more than 32
comparing to when it is recovered. This is expected that page
header/trailer data might differ.

But why did the page lsn on page#5 differed for page-size 4K only ?

We inserted the 16000 bytes in the blob. That means data was spread to
4 lob data pages.

(16000(data length) - 3392(data added to the first lob page)) /
(4039(data added to lob data pages)) = ~ 4 pages

For instance : Had we inserted the 261888 bytes then test would have
failed for all page sizes.

Fix
====
- Test file now has two scenarios :

  + Test#1 : Inserting the blob value such that data is *not* spread in
  number of lob data pages in multiple of commit frequency(i.e. 4) for
  all supported page sizes.

  + Test#2 : Inserting the blob value such that data is spread in number
  of lob data pages in multiple of commit frequency(i.e. 4) for
  all supported page sizes.

- Renamed the test to tablespace_recovery_state. I think changed name
  conveys the intent better. Did some minor improvement in the test file

- Optionally, introduced a test script too that let us compare the two
  tablespaces data either of all page or some page range. Used the
  script in the test to compare the data.

Change-Id: I0b5b0df0cd7e2d43adf0a8e3f6c83f1a43623a47