Bug#35750394 Use refactored TrpId-states to simplify ::connect_server
TransporterRegistry::connect_server() handled connect of a
multi transporter instance as a special case of a Transporter connect().

With refactorings in previous patches a multi Transporter instance
now has its own states++ available, and in most aspects are just a
plain Transporter, which enable some cleanup and simplifications.

In particular:

  - For a non-multi Transporter we used to check that performStates[]
    was 'CONNECTING'.
  - As a multi Transported instance used to not have its own
    performStates[], we instead checked !isConnected().

This patch introduce the utility method get_node_transporter_instance
which is now used to locate the Transporter instance to connect.

The performStates[TrpId] is used to check for CONNECTING state,
independent of whether a multi Transporter is used or not.

-> All special handling of a multi Transporter is removed.
... A multi Transporter instance is just a Transporter.

Change-Id: I4ae73b89846206e7afa499ff340930a230cf2745