Bug#35938286 Prepare TransporterRegistry for a Multi_Transporter not being a Transporter any more
As the Multi-Transporter is not a sub class of Transporter any more,
it can not be stored in theNodeIdTransporters[] any more.

-> theNodeIdTransporters[] is enhanced to always store the initial
base transporter to a specific NodeId. This also enables use to
simplify the implementation of ::get_node_base_transporter() as we
not need to handle that that theNodeIdTransporters[] may be a
Multi_Transporter any more.

In case a Multi_Transporter is created for a NodeId, it is now
stored in theNodeIdMultiTransporters[] instead. This array replace
theMultiTransporters[] with a slightly changes semantic - indexed
by NodeId instead, instead of 0..nMultiTransporters.

::get_node_multi_transporter() is also enhanced to look up
the Multi_Transporter directly from theNodeIdMultiTransporters[].

Change-Id: I8d9ad9dde262380da8dc5555a310eeffbdf0c487
(cherry picked from commit cb6f19b151cb5f2731e43874d1804d7a3856656b)