BUG#35042876: `group_replication_set_as_primary(member_uuid, timeout)` does not terminate DDLs
Details:
`group_replication_set_as_primary(member_uuid, timeout)` appoints a specific
member of the group as the new primary, overriding any election process.

The timeout parameter is an integer specifying a timeout in seconds for
transactions that are running on the existing primary when you use the
function.
When the timeout expires, for any transactions that did not yet reach their
commit phase, the client session is disconnected so that the transaction does
not proceed.

Problem:
A bug 'BUG#34664197: `group_replication_set_as_primary()` does not wait for
ongoing ALTER TABLE'was filed to make sure DDL are also killed when the timeout
option expires for the UDF group_replication_set_as_primary.
The related code changes were done under the 'WL#15497: Add statement type DDL
to transaction tracking facility'. The bug was filed to check the changes needed
in the group replication.

Resolution:
No code changes are needed from Group replication side.
Test case has been written to assert the timeout does terminates the DDLs.

Change-Id: I584a56e4848f425c699f54b2600dfeffeb188c52