Bug#35919302 pod deployment error 2336 Too many in Long Time Queue NDBD_EXIT_TIME_QUEUE_LONG
Problem:
IN SPJ there is a function, releaseGlobal, responsible to release
the excess of pages maintained in the m_free_page_list.
The  function iterates over the list, releases the objects and after
16 iterations it takes a real time break. In 'parallel' with the
real time break a new 'thread' of the releaseGlobal is spawned
(A CONTINUEB signal is sent with delay to self in order to start
the new releaseGlobal 'thread'). This can lead to an 'overflow'
of the Long Time Queue since there is no control over the number
of signal with delay that are being send.

Solution:
Do not send the extra delayed CONTINUEB creating a new 'thread'
when a real time break is taken.

Change-Id: Ifd6729a3412687f36dd29142f12a99682c1a9d22