Bug#35898221 Hypergraph: too low row estimate for semijoin that is not an equijoin
This commit fixes a regression introduced by the fix for bug
bug#34764211 "Too high row estimate for semijoin" (Change-Id:
I231cd0c8ef504d64cd835184a39f9975066b61bf). Some semijoins may be
transformed into an inner join between the right hand side aggregated
on the join fields, and the original left hand side. That fix
utilized this transform to make better row estimates, by estimating the
number of output rows as:

CARD(left_hand_relation) * inner_join_selectivity * CARD(d)

where:

* inner_join_selectivity is the cardinality of an inner join on the same
  predicate, divided by the cardinality of a cross join.

* 'd' is the set of distinct rows from right_hand_relation, when only
  looking at those columns that appear in the join predicate.

The regression happens for a semijoin that is not a pure
equijoin. That is, the join predicate is something else than a
conjunction of 'left_table.field=right_table.field' terms. In this
case JoinPredicate.semijoin_group is empty, because the semijoin to inner
join transform will never happen. In this case the
number of aggregated rows was wrongly set to 1. This fix corrects that.

This fix adds a new function EstimateSemijoinFanOut() that collects the
fields from the right hand_relation that appear in the join predicate.
Unlike JoinPredicate.semijoin_group, this works for both semijoin and
antijoin, and for arbitrary predicates, not just conjunctions of
field=field. That function then estimates CARD(d), using the same apparatus
that we use for e.g. DISTINCT.

Change-Id: I02224ff4f64315f2b0b92d6b883fe08e2f6a9975