Bug#35390341 Assertion `m_count > 0 && m_count > m_frame_null_count' failed.
This issue involves setting a user variable inside the argument of a
window function, which in turn is evaluated using the window frame
buffer (Window::m_needs_frame_buffering), row optimizable
(Window::m_row_optimizable).

Setting of user variables inside expressions is deprecated, in part
since the semantics are hard to define, but this patch avoids the
assert at least.

Normally when evaluating a function argument for a window function,
e.g.  AVG( 2 * c1 ), the function (here multiplication) will be
evaluated just before writing a row to the frame buffer, whereas
functions *containing* a window function will be evaluated later,
i.e. after the value of the window function is ready,
cf. split_sum_func2.

Now, for the first case when creating the window's output and frame
buffer tmp tables, we replace the function (here the multiplication)
with a result field in the window's tmp table,
cf. change_to_use_tmp_fields, so that the window function when
evaluated will pick up its input (the result of the multiplication)
from the window frame buffer. However, there is an explicit exception
for setting of user variables (which is implemented as a function),
cf. this comment in change_to_use_tmp_fields:

  /*
    Replace "@:=<expression>" with "@:=<tmp table column>". Otherwise, we
    would re-evaluate <expression>, and if expression were a subquery,
    this would access already-unlocked tables.
  */

which stems from Bug#11764371 at commit 6a2402b, which involves
locking issues when executing a subquery when evaluating a setting of
a user variable, a join and group by. The exception mentioned is a
problem for evaluating window functions, because the window function
(AVG) will try to evaluate Item_func_set_user_var when reading from
the frame buffer. But the underlying column is taken from the input
file, not from the output/frame buffer (it is not replaced by an
Item_ref subject to the slice indirection).

In the repro, the two first rows are null. The third row is non-null.
When trying to evaluate the wf for the third row, we invert the first
(null) row, but this time, since the third input row has been read,
the Item_func_set_user_var's argument points into the input buffer
which has a non-null row, where as AVG expects to invert a null row
(it has counters indicating that rows 1 and 2 are both null) from the
frame buffer. This causes the assert error seen.

The fix is to *not* apply the mentioned exception for setting user
variables when windowing: that way the user variable is only set when
we read the input result set for the windowing action, and the frame
buffer will contain that value. As far as the locking issue of
11764371, I tried windowing on the query mentioned for the issue
(included in test), and saw no issue.

Change-Id: I75dc538bea55d5888783586d31ced5e9643d6fa3